[Serve] Cleanup ThreadProxiedRouter; Strictly scope endpoint to router
